requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
tqdm>=4.62.0
PyYAML>=6.0
matplotlib>=3.5.0
